requirements.txt
lxml==4.9.3
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
python-docx==1.0.1
typing_extensions==4.8.0
